Added repository, coursesDao, viewmodel and temp layout for course selection. Course import works, TODO rest of the importation

+++ app/src/main/java/com/example/birdiediscgolf/Courses.kt
@@ -12,7 +12,7 @@ data class Course (
 
     val name: String,
 
-    val createdAt: Int
+    val createdAt: Long
     )
 
 @Entity(tableName = "holes")
@@ -22,7 +22,7 @@ data class Hole (
 
     val courseUuid: String,
 
-    val createdAt: Int,
+    val createdAt: Long,
 
     val hole: Int,
 
@@ -30,5 +30,5 @@ data class Hole (
 
     val par: Int,
 
-    val updatedAt: Int?
+    val updatedAt: Long?
 )
